back up music before update

+++ install.sh
@@ -2,8 +2,17 @@
 
 echo "installing paperbenni's dmenu menus"
 cd
-rm -rf paperbenni
-mkdir paperbenni
+
+# keep downloaded music
+if [ -e paperbenni]; then
+    if [ -e paperbenni/music]; then
+        echo "backing up music"
+        mv paperbenni/music .cache/papermusic
+    fi
+    rm -rf paperbenni
+    mkdir paperbenni
+fi
+
 cd paperbenni
 mkdir screenshots
 mkdir recordings music
@@ -19,3 +28,10 @@ rm -rf .git
 rm *.md
 rm install.sh
 chmod +x dm/*.sh
+
+cd
+
+if [ -e .papermusic ]; then
+    echo "restoring music"
+    mv .papermusic paperbenni/music
+fi
